Moved one prettyPrint function
